fix: adapt collision-resistant run id to current command.rs structure, split tests
